Features/employee hedy

+++ src/stores/MainStore.js
@@ -269,6 +269,11 @@ export const actions = store => ({
     axios(req)
       .then(response => {
         getOutlet(state.baseUrl, state.nameOutlet);
+        Swal.fire({
+          title: "Berhasil Menambahkan Outlet!",
+          icon: "success",
+          timer: 2000,
+        });
       })
       .catch(error => {});
   },
@@ -294,24 +299,47 @@ export const actions = store => ({
     axios(req)
       .then(response => {
         getOutlet(state.baseUrl, state.nameOutlet);
+        Swal.fire({
+          title: "Data Berhasil Diperbarui!",
+          icon: "success",
+          timer: 2000,
+        });
       })
       .catch(error => {});
   },
   deleteOutletById: (state, id) => {
-    const req = {
-      method: "delete",
-      url: `${state.baseUrl}/outlet/${id}`,
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${localStorage.getItem("token")}`
+    Swal.fire({
+      title: 'Yakin ?',
+      text: "Kamu tidak dapat mengembalikan data ini!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Ya, Hapus!',
+      cancelButtonText: 'Batal',
+    }).then((result) => {
+      if (result.value) {
+        const req = {
+          method: "delete",
+          url: `${state.baseUrl}/outlet/${id}`,
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: `Bearer ${localStorage.getItem("token")}`
+          }
+        };
+    
+        axios(req)
+          .then(response => {
+            getOutlet(state.baseUrl, state.nameOutlet);
+            Swal.fire(
+              'Terhapus!',
+              'File Anda telah dihapus.',
+              'success'
+            )
+          })
+          .catch(error => {});
       }
-    };
-
-    axios(req)
-      .then(response => {
-        getOutlet(state.baseUrl, state.nameOutlet);
-      })
-      .catch(error => {});
+    })
   },
   getOutletById: (state, id) => {
     const req = {
@@ -435,6 +463,11 @@ export const actions = store => ({
                 state.showProduct
               );
               getCategory(state.baseUrl);
+              Swal.fire({
+                title: "Berhasil Menambahkan Produk!",
+                icon: "success",
+                timer: 2000,
+              });
             })
             .catch(error => {
               store.setState({
@@ -573,6 +606,11 @@ export const actions = store => ({
                   state.showProduct
                 );
                 getCategory(state.baseUrl);
+                Swal.fire({
+                  title: "Data Berhasil Diperbarui!",
+                  icon: "success",
+                  timer: 2000,
+                });
               })
               .catch(error => {
                 store.setState({
@@ -608,45 +646,81 @@ export const actions = store => ({
   },
 
   deleteProduct: state => {
-    const req = {
-      method: "delete",
-      url: `${state.baseUrl}/product/delete/${state.idProduct}`,
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${localStorage.getItem("token")}`
-      }
-    };
-    axios(req)
-      .then(response => {
-        getProduct(
-          state.baseUrl,
-          state.category,
-          state.nameProduct,
-          state.showProduct
-        );
-        getCategory(state.baseUrl);
+    Swal.fire({
+      title: 'Yakin ?',
+      text: "Kamu tidak dapat mengembalikan data ini!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Ya, Hapus!',
+      cancelButtonText: 'Batal',
+    }).then((result) => {
+      if (result.value) {
+        const req = {
+          method: "delete",
+          url: `${state.baseUrl}/product/delete/${state.idProduct}`,
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: `Bearer ${localStorage.getItem("token")}`
+          }
+        };
+        axios(req)
+          .then(response => {
+            getProduct(
+              state.baseUrl,
+              state.category,
+              state.nameProduct,
+              state.showProduct
+            );
+            getCategory(state.baseUrl);
+            Swal.fire(
+              'Terhapus!',
+              'File Anda telah dihapus.',
+              'success'
+            )
+          })
+          .catch(error => {});
+        }
       })
-      .catch(error => {});
   },
   deleteProductById: (state, id) => {
-    const req = {
-      method: "delete",
-      url: `${state.baseUrl}/product/${id}`,
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${localStorage.getItem("token")}`
-      }
-    };
-    axios(req)
-      .then(response => {
-        getProduct(
-          state.baseUrl,
-          state.category,
-          state.nameProduct,
-          state.showProduct
-        );
+    Swal.fire({
+      title: 'Yakin ?',
+      text: "Kamu tidak dapat mengembalikan data ini!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Ya, Hapus!',
+      cancelButtonText: 'Batal',
+    }).then((result) => {
+      if (result.value) {
+        const req = {
+          method: "delete",
+          url: `${state.baseUrl}/product/${id}`,
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: `Bearer ${localStorage.getItem("token")}`
+          }
+        };
+        axios(req)
+          .then(response => {
+            getProduct(
+              state.baseUrl,
+              state.category,
+              state.nameProduct,
+              state.showProduct
+            );
+            Swal.fire(
+              'Terhapus!',
+              'File Anda telah dihapus.',
+              'success'
+            )
+          })
+          .catch(error => {});
+        }
       })
-      .catch(error => {});
   },
   getProductById: async (state, id) => {
     await localStorage.setItem("idProduct", id);
@@ -751,6 +825,11 @@ export const actions = store => ({
           state.nameInventory,
           state.statusInventory
         );
+        Swal.fire({
+          title: "Berhasil Menambahkan Bahan Baku!",
+          icon: "success",
+          timer: 2000,
+        });
       })
       .catch(error => {});
   },
@@ -778,6 +857,11 @@ export const actions = store => ({
           state.nameInventory,
           state.statusInventory
         );
+        Swal.fire({
+          title: "Data Berhasil Diperbarui!",
+          icon: "success",
+          timer: 2000,
+        });        
       })
       .catch(error => {});
   },
@@ -803,24 +887,42 @@ export const actions = store => ({
       .catch(error => {});
   },
   deleteInventoryById: async (state, id) => {
-    const req = {
-      method: "delete",
-      url: `${state.baseUrl}/inventory/detail/${id}`,
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${localStorage.getItem("token")}`
-      }
-    };
-    await axios(req)
-      .then(response => {
-        getInventory(
-          state.baseUrl,
-          state.outlet,
-          state.nameInventory,
-          state.statusInventory
-        );
+    Swal.fire({
+      title: 'Yakin ?',
+      text: "Kamu tidak dapat mengembalikan data ini!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Ya, Hapus!',
+      cancelButtonText: 'Batal',
+    }).then( async (result) => {
+      if (result.value) {
+        const req = {
+          method: "delete",
+          url: `${state.baseUrl}/inventory/detail/${id}`,
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: `Bearer ${localStorage.getItem("token")}`
+          }
+        };
+        await axios(req)
+          .then(response => {
+            getInventory(
+              state.baseUrl,
+              state.outlet,
+              state.nameInventory,
+              state.statusInventory
+            );
+            Swal.fire(
+              'Terhapus!',
+              'File Anda telah dihapus.',
+              'success'
+            )
+          })
+          .catch(error => {});
+        }
       })
-      .catch(error => {});
   },
   addStock: async state => {
     const req = {
@@ -843,6 +945,11 @@ export const actions = store => ({
           state.nameInventory,
           state.statusInventory
         );
+        Swal.fire({
+          title: "Berhasil Menambahkan Produk!",
+          icon: "success",
+          timer: 2000,
+        });
       })
       .catch(error => {});
   },
@@ -880,21 +987,38 @@ export const actions = store => ({
       .catch(error => {});
   },
   deleteCustomerById: (state, id) => {
-    store.setState({ isLoadingCustomer: true });
+    Swal.fire({
+      title: 'Yakin ?',
+      text: "Kamu tidak dapat mengembalikan data ini!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Ya, Hapus!',
+      cancelButtonText: 'Batal',
+    }).then((result) => {
+      if (result.value) {
+        store.setState({ isLoadingCustomer: true });
 
-    const req = {
-      method: "delete",
-      url: `${state.baseUrl}/customer/${id}`,
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${localStorage.getItem("token")}`
-      }
-    };
-    axios(req)
-      .then(response => {
-        getCustomer(state.baseUrl, state.nameCustomer);
+        const req = {
+          method: "delete",
+          url: `${state.baseUrl}/customer/${id}`,
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: `Bearer ${localStorage.getItem("token")}`
+          }
+        };
+        axios(req)
+          .then(response => {
+            Swal.fire(
+              'Terhapus!',
+              'File Anda telah dihapus.',
+              'success'
+            )
+          })
+          .catch(error => {});
+        }
       })
-      .catch(error => {});
   },
   getProvince: state => {
     const req = {
@@ -981,6 +1105,11 @@ export const actions = store => ({
           state.outlet,
           state.position
         );
+        Swal.fire({
+          text: "Berhasil Menambahkan Karyawan!",
+          icon: "success",
+          timer: 2000,
+        });
       })
       .catch(error => {});
   },
@@ -1007,24 +1136,42 @@ export const actions = store => ({
       .catch(error => {});
   },
   deleteEmployeeById: (state, id) => {
-    const req = {
-      method: "delete",
-      url: `${state.baseUrl}/employee/${id}`,
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${localStorage.getItem("token")}`
-      }
-    };
-    axios(req)
-      .then(response => {
-        getEmployee(
-          state.baseUrl,
-          state.nameEmployee,
-          state.outlet,
-          state.position
-        );
+    Swal.fire({
+      title: 'Yakin ?',
+      text: "Kamu tidak dapat mengembalikan data ini!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Ya, Hapus!',
+      cancelButtonText: 'Batal',
+    }).then((result) => {
+      if (result.value) {
+        const req = {
+          method: "delete",
+          url: `${state.baseUrl}/employee/${id}`,
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: `Bearer ${localStorage.getItem("token")}`
+          }
+        };
+        axios(req)
+          .then(response => {
+            getEmployee(
+              state.baseUrl,
+              state.nameEmployee,
+              state.outlet,
+              state.position
+            );
+            Swal.fire(
+              'Terhapus!',
+              'File Anda telah dihapus.',
+              'success'
+            )
+          })
+          .catch(error => {});
+        }
       })
-      .catch(error => {});
   },
   editEmployee: state => {
     const req = {
@@ -1050,6 +1197,11 @@ export const actions = store => ({
           state.outlet,
           state.position
         );
+        Swal.fire({
+          text: "Data Berhasil Diperbarui!",
+          icon: "success",
+          timer: 2000,
+        });
       })
       .catch(error => {});
   },
